requirements.txt
numpy
pillow
matplotlib
scikit-image
tensorflow
flask
opencv-python-headless